requirements.txt
PyPDF2
python-pptx
openai
streamlit
pytest
Pillow
gTTS==2.3.1